wallet create adding and debit is done need to test race condition

+++ backend/src/main/java/com/mohan/stock_pilot/wallet/dto/PaymentRequestDto.java
@@ -1,7 +1,7 @@
 package com.mohan.stock_pilot.wallet.dto;
 
-import java.math.BigDecimal;
+
 import java.util.UUID;
 
-public record PaymentRequestDto(UUID userId, BigDecimal amount, String paymentId) {
+public record PaymentRequestDto(UUID userId, long amount, String paymentId) {
 }
